python-service/requirements.txt
grpcio
grpcio-tools
scikit-learn
numpy
psycopg2-binary
